+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.10.0-rc.12",
-  genVersion: "2.658.3",
-  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.12 2.658.3 2.0 @orq-ai/node",
+  sdkVersion: "3.10.0-rc.13",
+  genVersion: "2.660.0",
+  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.13 2.660.0 2.0 @orq-ai/node",
 } as const;