+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -262,6 +262,7 @@ export const GetAllPromptsProvider = {
   Elevenlabs: "elevenlabs",
   Litellm: "litellm",
   Openailike: "openailike",
+  Cerebras: "cerebras",
 } as const;
 export type GetAllPromptsProvider = ClosedEnum<typeof GetAllPromptsProvider>;