+++ packages/orq-rc/RELEASES.md
@@ -2559,3 +2559,13 @@ Based on:
 - [typescript v3.10.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.12 - packages/orq-rc
+
+## 2025-07-22 06:23:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.584.0 (2.660.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.13 - packages/orq-rc